backend/requirements.txt
Flask
werkzeug
youtube-transcript-api
python-dotenv